server/source/requirements.txt
googlemaps
fastapi
uvicorn
PyMySQL
cloud-sql-python-connector
mysql.connector
pandas